Fail fast in N-R when NaN or Infinity is encountered
Encapsulate the state of the N-R calculation in a state class.  Use
the setters on the Calculation class to ensure that an exception is
thrown when a value is NaN or Infinity.

Add parent class exception to ZeroValuedDerivativeException to represent
when a calculation overflow has occurred.  The OverflowException 
allows access to the data of the Calculation state class for N-R.

Fixes #9